Add I/O facilities for the protobuf wire format.
This adds two things:

InputStreamBuffer and OutputStreamBuffer, which keep track of a window
of bytes in memory belonging to a stream of input or output data.
Subclasses fill in the specifics of where the buffers come from or go
to.

ProtoReader and ProtoWriter can be used to read and write data in
protobuf wire format encoding from an underlying stream buffer.

BUG: 25762536
Change-Id: Ia76bee256d90c7fcf4c062b1f448d91b3afe8d81