Basic storage interface for NVRAM implementation.
This lays down the basic interfaces for reading and writing persistent
storage objects used by the NVRAM implementation. We'll have two
different types of objects: The header contains global bookkeeping
information as well as list of spaces that are currently defined. For
each space, there's a space-specific storage object that holds space
metadata and contents.

BUG: 25762540
Change-Id: Iadce20f76d84b3ed1ea25762ec673daae31d6b1d