Add a generic Option type.
Optional is a class template that implements the generic Option type,
i.e. holding a value of arbitrary type along with a flag indicating
validity. This change also adjusts the protobuf encoder so Optional is
encoded and decoded according to the semantics for optional protobuf
message fields. Note that Optional is only required for protobuf
encoding if the code processing the message doesn't want to serialize
absent optional fields or needs to retain information about whether an
optional field was present in a decoded message.

BUG: 25762536
Change-Id: I1dfff56224226575775c64911ab74c2e99c3ec09