Implement NVRAM HAL glue code.
This adds glue code to simplify the creation of NVRAM HAL modules. It
bridges the NVRAM HAL C API surface to wrap invocation parameters in
an |nvram::Request| object, dispatches the operation to a delegate,
which will fill return the result in an |nvram::Response| object that
is then decoded to complete the HAL function invocation.

BUG: 26876961
Change-Id: I7bc5cd31ed43430865cc60105740eab17367ca90