Fix integer overflow on large buffer reads.
This fixes a problem where NestedInputStreamBuffer would fail to
detect integer overflow on large read requests, and hitting an
NVRAM_CHECK() assertion subsequently. This proves that the I/O code is
subtle enough to warrant better test coverage, so this change adds
unit tests for the byte-level I/O code.

Bug: 27447078
Change-Id: I3dbbcd731ec3c501baf1d637aa5d97e0756449b5