Isolate crypto implementation details.
This moves the details of interfacing with crypto libraries behind a
clean interface, so crypto can be swapped out as needed. Now that
swapping out crypto libraries is easy, use BoringSSL for the main
Android build. We can now always replace it with something else if
that's need in other environments.

BUG: 29238681

Change-Id: I9af046b62273ede4c98d4239a4774c25764dbdaf