Add commands for clearing NVRAM on full reset.
This adds commands that allow clearing the entire NVRAM contents. This
can be used by firmware to implement the correct behavior as part of
device recovery. There's always the risk of malicious invocation of
the reset logic, so the new code is behind a compile flag which isn't
enabled by default.

BUG: 29260086
Change-Id: I8eb8a4953ed62ff75b8242daa60c1e4f99bf254d